bus/hp_dio/98644: convert to anonymous namespace (nw)
