+++ src/devices/bus/hp_dio/hp98644.cpp
@@ -9,15 +9,44 @@
 #include "emu.h"
 #include "hp98644.h"
 #include "bus/rs232/rs232.h"
+#include "machine/ins8250.h"
+#include "hp_dio.h"
+
+namespace {
+
+class dio16_98644_device :
+		public device_t,
+		public bus::hp_dio::device_dio16_card_interface
+{
+public:
+	// construction/destruction
+	dio16_98644_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock);
+
+protected:
+	dio16_98644_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, uint32_t clock);
 
 
-//**************************************************************************
-//  GLOBAL VARIABLES
-//**************************************************************************
+	// device-level overrides
+	virtual void device_start() override;
+	virtual void device_reset() override;
 
-DEFINE_DEVICE_TYPE(HPDIO_98644, bus::hp_dio::dio16_98644_device, "dio98644", "HP98644A Asynchronous Serial Interface")
+	virtual ioport_constructor device_input_ports() const override;
+	// optional information overrides
+	virtual void device_add_mconfig(machine_config &config) override;
 
-namespace bus::hp_dio {
+	uint16_t io_r(offs_t offset);
+	void io_w(offs_t offset, uint16_t data);
+
+	required_device<ins8250_device> m_uart;
+
+private:
+	required_ioport m_switches;
+	bool     m_installed_io;
+	uint8_t  m_control;
+
+	bool     m_loopback;
+	uint8_t  m_data;
+};
 
 //-------------------------------------------------
 //  device_add_mconfig - add device configuration
@@ -240,4 +269,7 @@ void dio16_98644_device::io_w(offs_t offset, uint16_t data)
 	}
 }
 
-} // namespace bus::hp_dio
+} // anonymous namespace
+
+DEFINE_DEVICE_TYPE_PRIVATE(HPDIO_98644, bus::hp_dio::device_dio16_card_interface, dio16_98644_device, "dio98644", "HP98644A Asynchronous Serial Interface")
+
